pyproject.toml
[tool.poetry]
name = "telegram_aiml_bot"
version = "0.1.0"
description = ""
authors = ["Yusuf Kaka <yusufk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-aiml = "^0.9.3"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

